перевел на тайпскрипт, пофиксил изменение профиля
