+++ .gitignore
@@ -217,16 +217,16 @@ typings/
 .cache
 .parcel-cache
 
-# Next.js build output
+# Next.ts build output
 .next
 
-# Nuxt.js build / generate output
+# Nuxt.ts build / generate output
 .nuxt
 dist
 
 # Gatsby files
 .cache/
-# Comment in the public line in if your project uses Gatsby and not Next.js
+# Comment in the public line in if your project uses Gatsby and not Next.ts
 # https://nextjs.org/blog/next-9-1#public-directory-support
 # public
