+++ package.json
@@ -6,7 +6,6 @@
   "scripts": {
     "start": "DEBUG=* node server",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "pug-build": "./pug.sh",
     "install": "",
     "build": "webpack"
   },
@@ -28,6 +27,7 @@
     "node-fetch": "^2.6.1"
   },
   "devDependencies": {
+    "@types/node": "^14.14.37",
     "autoprefixer": "^10.2.5",
     "css-loader": "^5.2.0",
     "eslint": "^7.23.0",
@@ -45,6 +45,8 @@
     "sass": "^1.32.8",
     "sass-loader": "^11.0.1",
     "style-loader": "^2.0.0",
+    "ts-loader": "^8.1.0",
+    "typescript": "^4.2.4",
     "webpack": "^5.31.0",
     "webpack-cli": "^4.6.0"
   }
