--- pug.sh
@@ -1,15 +1,0 @@
-#!/bin/sh
-(pug -c -E tmpl -n oneTableEventTemplate public/templates/events/one-table-event.pug)
-(pug -c -E tmpl -n oneEventPageTemplate public/templates/one-event-page/one-event-page.pug)
-(pug -c -E tmpl -n loginTemplate public/templates/login/login.pug)
-(pug -c -E tmpl -n signUpFormTemplate public/templates/signup/signup.pug)
-(pug -c -E tmpl -n upperTextTemplate public/templates/events/upper-text.pug)
-(pug -c -E tmpl -n navbarTemplate public/components/navbar/navbar.pug)
-(pug -c -E tmpl -n navbarLoggedTemplate public/components/navbar/navbar-logged.pug)
-(pug -c -E tmpl -n profileTemplate public/templates/profile/profile.pug)
-(pug -c -E tmpl -n myProfileTemplate public/templates/my-profile/my-profile.pug)
-(pug -c -E tmpl -n myProfileAboutTabTemplate public/templates/my-profile-about-tab/my-profile-about-tab.pug)
-(pug -c -E tmpl -n myProfileSettingsTabTemplate public/templates/my-profile-settings-tab/my-profile-settings-tab.pug)
-(pug -c -E tmpl -n myProfileEventsTabTemplate public/templates/my-profile-events-tab/my-profile-events-tab.pug)
-(pug -c -E tmpl -n errorLoginTemplate public/templates/login/error-login.pug)
-(pug -c -E tmpl -n errorSignupTemplate public/templates/signup/errorSignup.pug)
