+++ src/index.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="viewport" content="width=device-width">
+    <title>QDA</title>
 </head>
 <body>
     <div id="navbar" class="navbar"></div>
