--- src/js/dispatcher/dispatcher.js
@@ -1,13 +1,0 @@
-export default class Dispatcher {
-  constructor() {
-    this.channel = [];
-  }
-
-  dispatch(action) {
-    this.channel.forEach((listener) => listener(action));
-  }
-
-  register(listener) {
-    this.channel.push(listener);
-  }
-}
