+++ src/ts/dispatcher/dispatcher.ts
@@ -1,0 +1,17 @@
+import {ActionsInterface} from "../actions/actions";
+
+export default class Dispatcher {
+  public channel: Array<Function>;
+
+  constructor() {
+    this.channel = [];
+  }
+
+  dispatch(action: ActionsInterface) {
+    this.channel.forEach((listener) => listener(action));
+  }
+
+  register(listener: Function) {
+    this.channel.push(listener);
+  }
+}
