+++ src/ts/storage/EventsStore.ts
@@ -1,0 +1,33 @@
+import { getAllEventsJson } from '../networkModule/network';
+import { channelNames } from '../config/config';
+import Store from "./store";
+import Actions, {ActionsInterface} from "../actions/actions";
+
+const allEventsSymbol = Symbol('allEventsSymbol');
+const globalStoreSymbol = Symbol('globalStoreSymbol');
+
+export default class EventsStore {
+  public globalStore: Store;
+  public allEvents: object;
+
+  constructor(globalStore: Store) {
+    this.globalStore = globalStore;
+    this.allEvents = {};
+  }
+
+  async update() {
+    this.allEvents = await getAllEventsJson();
+    this.globalStore.eventBus.publish(channelNames.eventsUpdated);
+  }
+
+  reducer(action: ActionsInterface) {
+    switch (action.eventName) {
+      case 'events/update':
+        this.update();
+        break;
+
+      default:
+        break;
+    }
+  }
+}
